New quest system - NPCs want you to find other NPCs
Just ask for a `quest` and the NPC will ask you to find someone from any
city. The reward is based on the distance from the city. However, no
currency system is implemented yet, so no reward is actually awarded.
This "quest system" still needs a ton of work and probably a rewrite.